requirements.txt
python-dotenv==1.0.1
langchain-community==0.3.2
langchain-openai==0.2.3
langchain-groq==0.2.0
langchainhub==0.1.21
pypdf==4.2.0
faiss-cpu==1.9.0
langchain-huggingface==0.1.0
streamlit==1.40.0
fpdf==1.7.2
bs4==0.0.2
